Fix account removal, #6012
